pyproject.toml

[tool.poetry]
name = "project-a-assistant"
version = "0.1.0"
description = "Conversational assistant for B2B sales teams"
authors = ["Generated <assistant@example.com>"]
packages = [{ include = "project_a_assistant", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.4.5"
fastapi = "^0.111"
uvicorn = { extras = ["standard"], version = "^0.29" }
duckduckgo-search = "^5.3"
httpx = "^0.27"
python-dotenv = "^1.0"
pydantic = "^2.7"
pydantic-settings = "^2.9.1"
python-multipart = "^0.0.20"
langchain = "^0.3.25"
tiktoken = "^0.9.0"
langchain-openai = "^0.3.17"
langchain-community = "^0.3.24"
azure-storage-blob = "^12.25.1"
requests = "^2.32.3"
langchain-mcp-adapters = "^0.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
ruff = "^0.5.3"
pytest-benchmark = { version = "^4.0", optional = true }
pytest-asyncio = "^0.26.0"
langsmith = "^0.3"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
  "asyncio: mark a test as asyncio",
]
